Fixing gate issues

+++ app/Http/Controllers/Admin/NotesController.php
@@ -18,7 +18,9 @@ class NotesController extends Controller
      */
     public function index()
     {
-
+        if (! Gate::allows('note_access')) {
+            return abort(401);
+        }
 
         if (request('show_deleted') == 1) {
             if (! Gate::allows('note_delete')) {
@@ -39,21 +41,28 @@ class NotesController extends Controller
      */
     public function create()
     {
+        if (! Gate::allows('note_create')) {
+            return abort(401);
+        }
+
         $properties = \App\Property::get()->pluck('name', 'id');
+
         return view('admin.notes.create', compact('properties'));
     }
 
     /**
      * Store a newly created Note in storage.
      *
-     * @param  \App\Http\Requests\StoreNotesRequest  $request
+     * @param  \App\Http\Requests\StoreNotesRequest $request
      * @return \Illuminate\Http\Response
      */
     public function store(StoreNotesRequest $request)
     {
+        if (! Gate::allows('note_create')) {
+            return abort(401);
+        }
+
         $note = Note::create($request->all());
-
-
 
         return redirect()->route('admin.notes.index');
     }
@@ -62,29 +71,36 @@ class NotesController extends Controller
     /**
      * Show the form for editing Note.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
     {
+        if (! Gate::allows('note_edit')) {
+            return abort(401);
+        }
+
         $properties = \App\Property::get()->pluck('name', 'id');
-        $note = Note::findOrFail($id);
+        $note       = Note::findOrFail($id);
+
         return view('admin.notes.edit', compact('note', 'properties'));
     }
 
     /**
      * Update Note in storage.
      *
-     * @param  \App\Http\Requests\UpdateNotesRequest  $request
-     * @param  int  $id
+     * @param  \App\Http\Requests\UpdateNotesRequest $request
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function update(UpdateNotesRequest $request, $id)
     {
+        if (! Gate::allows('note_edit')) {
+            return abort(401);
+        }
+
         $note = Note::findOrFail($id);
         $note->update($request->all());
-
-
 
         return redirect()->route('admin.notes.index');
     }
@@ -93,11 +109,15 @@ class NotesController extends Controller
     /**
      * Display Note.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
     {
+        if (! Gate::allows('note_view')) {
+            return abort(401);
+        }
+
         $note = Note::findOrFail($id);
 
         return view('admin.notes.show', compact('note'));
@@ -107,11 +127,15 @@ class NotesController extends Controller
     /**
      * Remove Note from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
+        if (! Gate::allows('note_delete')) {
+            return abort(401);
+        }
+
         $note = Note::findOrFail($id);
         $note->delete();
 
@@ -125,6 +149,10 @@ class NotesController extends Controller
      */
     public function massDestroy(Request $request)
     {
+        if (! Gate::allows('note_delete')) {
+            return abort(401);
+        }
+
         if ($request->input('ids')) {
             $entries = Note::whereIn('id', $request->input('ids'))->get();
 
@@ -138,11 +166,15 @@ class NotesController extends Controller
     /**
      * Restore Note from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function restore($id)
     {
+        if (! Gate::allows('note_delete')) {
+            return abort(401);
+        }
+
         $note = Note::onlyTrashed()->findOrFail($id);
         $note->restore();
 
@@ -152,11 +184,15 @@ class NotesController extends Controller
     /**
      * Permanently delete Note from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function perma_del($id)
     {
+        if (! Gate::allows('note_delete')) {
+            return abort(401);
+        }
+
         $note = Note::onlyTrashed()->findOrFail($id);
         $note->forceDelete();
