+++ app/Http/Controllers/Admin/PermissionsController.php
@@ -18,9 +18,11 @@ class PermissionsController extends Controller
      */
     public function index()
     {
+        if (! Gate::allows('permission_access')) {
+            return abort(401);
+        }
 
-
-                $permissions = Permission::all();
+        $permissions = Permission::all();
 
         return view('admin.permissions.index', compact('permissions'));
     }
@@ -32,20 +34,26 @@ class PermissionsController extends Controller
      */
     public function create()
     {
+        if (! Gate::allows('permission_create')) {
+            return abort(401);
+        }
+
         return view('admin.permissions.create');
     }
 
     /**
      * Store a newly created Permission in storage.
      *
-     * @param  \App\Http\Requests\StorePermissionsRequest  $request
+     * @param  \App\Http\Requests\StorePermissionsRequest $request
      * @return \Illuminate\Http\Response
      */
     public function store(StorePermissionsRequest $request)
     {
+        if (! Gate::allows('permission_create')) {
+            return abort(401);
+        }
+
         $permission = Permission::create($request->all());
-
-
 
         return redirect()->route('admin.permissions.index');
     }
@@ -54,11 +62,15 @@ class PermissionsController extends Controller
     /**
      * Show the form for editing Permission.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
     {
+        if (! Gate::allows('permission_edit')) {
+            return abort(401);
+        }
+
         $permission = Permission::findOrFail($id);
 
         return view('admin.permissions.edit', compact('permission'));
@@ -67,16 +79,18 @@ class PermissionsController extends Controller
     /**
      * Update Permission in storage.
      *
-     * @param  \App\Http\Requests\UpdatePermissionsRequest  $request
-     * @param  int  $id
+     * @param  \App\Http\Requests\UpdatePermissionsRequest $request
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function update(UpdatePermissionsRequest $request, $id)
     {
+        if (! Gate::allows('permission_edit')) {
+            return abort(401);
+        }
+
         $permission = Permission::findOrFail($id);
         $permission->update($request->all());
-
-
 
         return redirect()->route('admin.permissions.index');
     }
@@ -85,15 +99,19 @@ class PermissionsController extends Controller
     /**
      * Display Permission.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
     {
+        if (! Gate::allows('permission_view')) {
+            return abort(401);
+        }
+
         $roles = \App\Role::whereHas('permission',
-                    function ($query) use ($id) {
-                        $query->where('id', $id);
-                    })->get();
+            function ($query) use ($id) {
+                $query->where('id', $id);
+            })->get();
 
         $permission = Permission::findOrFail($id);
 
@@ -104,11 +122,15 @@ class PermissionsController extends Controller
     /**
      * Remove Permission from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
+        if (! Gate::allows('permission_delete')) {
+            return abort(401);
+        }
+
         $permission = Permission::findOrFail($id);
         $permission->delete();
 
@@ -122,6 +144,10 @@ class PermissionsController extends Controller
      */
     public function massDestroy(Request $request)
     {
+        if (! Gate::allows('permission_delete')) {
+            return abort(401);
+        }
+
         if ($request->input('ids')) {
             $entries = Permission::whereIn('id', $request->input('ids'))->get();
