+++ app/Http/Controllers/Admin/RolesController.php
@@ -18,9 +18,11 @@ class RolesController extends Controller
      */
     public function index()
     {
+        if (! Gate::allows('role_access')) {
+            return abort(401);
+        }
 
-
-                $roles = Role::all();
+        $roles = Role::all();
 
         return view('admin.roles.index', compact('roles'));
     }
@@ -32,9 +34,11 @@ class RolesController extends Controller
      */
     public function create()
     {
-        
-        $permissions = \App\Permission::get()->pluck('title', 'id');
+        if (! Gate::allows('role_create')) {
+            return abort(401);
+        }
 
+        $permissions = \App\Permission::get()->pluck('title', 'id');
 
         return view('admin.roles.create', compact('permissions'));
     }
@@ -42,15 +46,17 @@ class RolesController extends Controller
     /**
      * Store a newly created Role in storage.
      *
-     * @param  \App\Http\Requests\StoreRolesRequest  $request
+     * @param  \App\Http\Requests\StoreRolesRequest $request
      * @return \Illuminate\Http\Response
      */
     public function store(StoreRolesRequest $request)
     {
+        if (! Gate::allows('role_create')) {
+            return abort(401);
+        }
+
         $role = Role::create($request->all());
         $role->permission()->sync(array_filter((array)$request->input('permission')));
-
-
 
         return redirect()->route('admin.roles.index');
     }
@@ -59,16 +65,17 @@ class RolesController extends Controller
     /**
      * Show the form for editing Role.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
     {
-        
+        if (! Gate::allows('role_edit')) {
+            return abort(401);
+        }
+
         $permissions = \App\Permission::get()->pluck('title', 'id');
-
-
-        $role = Role::findOrFail($id);
+        $role        = Role::findOrFail($id);
 
         return view('admin.roles.edit', compact('role', 'permissions'));
     }
@@ -76,17 +83,19 @@ class RolesController extends Controller
     /**
      * Update Role in storage.
      *
-     * @param  \App\Http\Requests\UpdateRolesRequest  $request
-     * @param  int  $id
+     * @param  \App\Http\Requests\UpdateRolesRequest $request
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function update(UpdateRolesRequest $request, $id)
     {
+        if (! Gate::allows('role_edit')) {
+            return abort(401);
+        }
+
         $role = Role::findOrFail($id);
         $role->update($request->all());
         $role->permission()->sync(array_filter((array)$request->input('permission')));
-
-
 
         return redirect()->route('admin.roles.index');
     }
@@ -95,19 +104,21 @@ class RolesController extends Controller
     /**
      * Display Role.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
     {
-        
-        $permissions = \App\Permission::get()->pluck('title', 'id');
-$users = \App\User::whereHas('role',
-                    function ($query) use ($id) {
-                        $query->where('id', $id);
-                    })->get();
+        if (! Gate::allows('role_view')) {
+            return abort(401);
+        }
 
-        $role = Role::findOrFail($id);
+        $permissions = \App\Permission::get()->pluck('title', 'id');
+        $users       = \App\User::whereHas('role',
+            function ($query) use ($id) {
+                $query->where('id', $id);
+            })->get();
+        $role        = Role::findOrFail($id);
 
         return view('admin.roles.show', compact('role', 'users'));
     }
@@ -116,11 +127,15 @@ $users = \App\User::whereHas('role',
     /**
      * Remove Role from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
+        if (! Gate::allows('role_delete')) {
+            return abort(401);
+        }
+
         $role = Role::findOrFail($id);
         $role->delete();
 
@@ -134,6 +149,10 @@ $users = \App\User::whereHas('role',
      */
     public function massDestroy(Request $request)
     {
+        if (! Gate::allows('role_delete')) {
+            return abort(401);
+        }
+
         if ($request->input('ids')) {
             $entries = Role::whereIn('id', $request->input('ids'))->get();
